Rename package to rstinder
