+++ rstfinder/collapse_rst_labels.py
