+++ rstfinder/utils/LICENSE_d3.txt
